refactor(comments): share Slack identity resolution via posthog.helpers
Move the token-agnostic Slack identity helpers (resolve_slack_user,
resolve_slack_avatar_by_email, resolve_posthog_user_for_slack) and their profile/avatar
Redis caches out of the conversations product into posthog/helpers/slack_identity.py so
core discussions can reuse them when mirroring comment threads to/from Slack. Each takes a
WebClient, so the caller supplies the workspace token. Bot-user-id resolution stays in
conversations (the @PostHog app has its own). Behavior unchanged; cache keys move to a
neutral slack_identity: namespace (one-time cold cache on deploy).

Generated-By: PostHog Code
Task-Id: 9ac92446-4916-4c82-80ba-7f554be8f479
