+++ products/conversations/backend/cache.py
@@ -25,7 +25,6 @@ logger = structlog.get_logger(__name__)
 MESSAGES_CACHE_TTL = 15  # Short TTL - messages need to appear quickly
 TICKETS_CACHE_TTL = 30  # Slightly longer - ticket list changes less frequently
 UNREAD_COUNT_CACHE_TTL = 30  # For dashboard polling - invalidated on changes
-SLACK_AVATAR_CACHE_TTL = 5 * 60  # 5 minutes for slack avatar lookup
 
 # All possible status filter values for tickets cache invalidation
 _TICKETS_STATUS_VARIANTS: list[str | None] = [None, *[s.value for s in Status]]
@@ -150,24 +149,6 @@ def invalidate_tickets_cache(team_id: int, widget_session_id: str) -> None:
         logger.warning("conversations_cache_invalidate_error", keys=keys)
 
 
-# Slack User Cache
-# Caches Slack user profile lookups (name, email, avatar) to reduce
-# Slack API calls. Keyed by slack_user_id — IDs are functionally unique
-# across workspaces. Short TTL keeps profiles reasonably fresh.
-
-SLACK_USER_CACHE_TTL = 5 * 60  # 5 minutes
-
-
-def get_cached_slack_user(slack_user_id: str) -> dict | None:
-    """Get cached Slack user profile."""
-    key = _make_cache_key("slack_user", slack_user_id)
-    try:
-        return cache.get(key)
-    except Exception:
-        logger.warning("conversations_cache_get_error", key=key)
-        return None
-
-
 PERSON_DISTINCT_IDS_CACHE_TTL = 30  # seconds
 
 
@@ -197,40 +178,6 @@ def get_person_distinct_ids(team_id: int, distinct_id: str) -> list[str]:
         logger.warning("conversations_cache_set_error", key=key)
 
     return all_ids
-
-
-def set_cached_slack_user(slack_user_id: str, user_info: dict) -> None:
-    """Cache a Slack user profile."""
-    key = _make_cache_key("slack_user", slack_user_id)
-    try:
-        cache.set(key, user_info, timeout=SLACK_USER_CACHE_TTL)
-    except Exception:
-        logger.warning("conversations_cache_set_error", key=key)
-
-
-# Slack Avatar-by-Email Cache
-# Caches the result of users.lookupByEmail so outbound replies can show
-# the replying user's Slack profile picture. Empty string = negative cache
-# (user not found in the Slack workspace for that email).
-
-
-def get_cached_slack_avatar(email: str) -> str | None:
-    """Get cached Slack avatar URL for an email. Returns None on cache miss, empty string for negative cache."""
-    key = _make_cache_key("slack_avatar", email.lower())
-    try:
-        return cache.get(key)
-    except Exception:
-        logger.warning("conversations_cache_get_error", key=key)
-        return None
-
-
-def set_cached_slack_avatar(email: str, avatar_url: str) -> None:
-    """Cache a Slack avatar URL (or empty string for negative cache)."""
-    key = _make_cache_key("slack_avatar", email.lower())
-    try:
-        cache.set(key, avatar_url, timeout=SLACK_AVATAR_CACHE_TTL)
-    except Exception:
-        logger.warning("conversations_cache_set_error", key=key)
 
 
 # Slack Bot User ID Cache
